solving code review issues
