--- solutions/devsprint-gabriel-goto-4/app/src/test/java/com/devpass/spaceapp/faker/RocketDetailsViewModelFake.kt
@@ -1,17 +1,0 @@
-package com.devpass.spaceapp.faker
-
-import com.devpass.spaceapp.model.RocketDetail
-
-class RocketDetailsViewModelFake {
-
-    object FakeRocketDetailsViewModel {
-
-        val fakeRocketDetail = RocketDetail(
-            id = "5e9d0d95eda69955f709d1eb",
-            name = "Falcon 1",
-            description = "The Falcon 1 was an expendable launch system privately developed and manufactured by SpaceX during 2006-2009. On 28 September 2008, Falcon 1 became the first privately-developed liquid-fuel launch vehicle to go into orbit around the Earth.",
-            image = "https://imgur.com/DaCfMsj.jpg"
-        )
-
-    }
-}
